refactor: QuickBookDialog → BookingCreateDialog, collapse dual dialogs
Renames everywhere to align with the actual architecture: BookingDialog
is the single booking-add dialog; its create branch delegates to a
BookingCreateDialog (the composed picker-sections flow). "Quick Book"
was a legacy label from when the dialog was a lightweight alternative to
a flat-form create — now that it IS the create workflow, keeping the
name was actively misleading.

File renames (git mv):
- packages/ui/registry/bookings/quick-book-dialog.tsx → booking-create-dialog.tsx
- templates/operator/src/components/voyant/bookings/quick-book-dialog.tsx → booking-create-dialog.tsx
- templates/dmc/src/components/voyant/bookings/quick-book-dialog.tsx → booking-create-dialog.tsx

Symbol renames:
- QuickBookDialog / QuickBookDialogProps → BookingCreateDialog / BookingCreateDialogProps

i18n (packages/i18n/src/admin/bookings.ts):
- bookings.list.quickBook label removed — booking-list now has a single
  "+ New Booking" button (en + ro).
- bookings.quickBook namespace → bookings.create (en + ro).

Registry (packages/ui/registry.json):
- voyant-bookings-quick-book-dialog → voyant-bookings-booking-create-dialog
  (path + target updated).
- voyant-bookings-booking-dialog now declares
  voyant-bookings-booking-create-dialog as a registry dep.
- voyant-bookings-booking-list description cleaned; dropped the
  booking-create-dialog dep (consumed transitively via booking-dialog).

Consumer updates:
- booking-list.tsx (registry, operator stayed pointing at op-specific,
  dmc): dropped the two-dialog state (dialogOpen + quickBookOpen) in
  favor of a single BookingDialog whose create/edit mode is driven by
  whether an editing booking was selected. Dropped the Zap icon import.
- product-detail-page.tsx (operator): local state renamed
  quickBookOpen → bookingCreateOpen.
- dmc BookingDialog upgraded to the same wrap-around pattern operator
  uses — create branch delegates to BookingCreateDialog, edit branch
  renders the flat form. Previously dmc had no such delegation so its
  legacy flat form did both; now dmc gets the composed picker flow for
  create for free.

Typecheck clean on i18n, operator, dmc.
