+++ packages/ui/registry/bookings/booking-create-dialog.tsx
@@ -31,7 +31,6 @@ import {
   SelectValue,
   Textarea,
 } from "@/components/ui"
-import { useAdminMessages } from "@/lib/admin-i18n"
 
 import {
   emptyPassengerListValue,
@@ -75,12 +74,6 @@ function generateBookingNumber(): string {
   return `BK-${y}${m}-${seq}`
 }
 
-/**
- * Convert a PaymentScheduleValue (section's UI shape) to the rows the
- * `/quick-create` endpoint expects. Returns an empty array for `unpaid` —
- * the orchestrator treats a zero-length paymentSchedules as "operator will
- * invoice manually."
- */
 function paymentScheduleToRows(
   value: PaymentScheduleValue,
   currency: string,
@@ -152,7 +145,7 @@ function passengersToRows(value: PassengerListValue): QuickCreateTravelerInput[]
     firstName: p.firstName.trim(),
     lastName: p.lastName.trim(),
     email: p.email.trim() || null,
-    participantType: p.role === "lead" || p.role === "adult" ? "traveler" : "traveler",
+    participantType: "traveler",
     travelerCategory:
       p.role === "child"
         ? "child"
@@ -166,7 +159,7 @@ function passengersToRows(value: PassengerListValue): QuickCreateTravelerInput[]
   }))
 }
 
-export interface QuickBookDialogProps {
+export interface BookingCreateDialogProps {
   open: boolean
   onOpenChange: (open: boolean) => void
   onCreated?: (booking: BookingRecord) => void
@@ -174,13 +167,23 @@ export interface QuickBookDialogProps {
   defaultProductId?: string
 }
 
-export function QuickBookDialog({
+/**
+ * Operator booking-create dialog. Composes the booking-create picker
+ * sections — product, departure, rooms, person, shared-room, passengers,
+ * price breakdown, voucher, payment schedule — and submits via the atomic
+ * `POST /v1/bookings/quick-create` endpoint so partial failures can't
+ * leave orphan state.
+ *
+ * Normally consumed via `BookingDialog` which delegates here when no
+ * `booking` prop is passed. Apps that need a bespoke flow can install the
+ * sections individually and assemble their own dialog instead of forking.
+ */
+export function BookingCreateDialog({
   open,
   onOpenChange,
   onCreated,
   defaultProductId,
-}: QuickBookDialogProps) {
-  const messages = useAdminMessages().bookings.quickBook
+}: BookingCreateDialogProps) {
   const [product, setProduct] = React.useState<ProductPickerValue>({
     productId: defaultProductId ?? "",
     optionId: null,
@@ -194,9 +197,13 @@ export function QuickBookDialog({
   const [paymentSchedule, setPaymentSchedule] =
     React.useState<PaymentScheduleValue>(emptyPaymentScheduleValue)
   const [notes, setNotes] = React.useState("")
-  // Post-create: confirm + fire auto-dispatch. The operator template wires
-  // `autoConfirmAndDispatch` on the notifications module so a booking.confirmed
-  // transition auto-sends the doc bundle — no separate notify call needed.
+  /**
+   * Optional post-create transition: set status to `confirmed` right after
+   * create succeeds. When the parent app has the notifications module's
+   * `autoConfirmAndDispatch` enabled, this fires the doc bundle + traveler
+   * email via the `booking.confirmed` subscriber. When it isn't, the
+   * booking simply lands in `confirmed` instead of `draft`.
+   */
   const [confirmAfterCreate, setConfirmAfterCreate] = React.useState(false)
   const [error, setError] = React.useState<string | null>(null)
 
@@ -256,11 +263,6 @@ export function QuickBookDialog({
     return `${date}${remaining}`
   }, [])
 
-  // Passenger room-assignment options. The capacity ceiling is
-  // `rooms[unit] × occupancyMax` seats; we subtract the count of passengers
-  // already assigned to each unit so the per-unit selector in the passenger
-  // row can disable full rooms. occupancyMax defaults to 1 when the unit
-  // definition doesn't carry it.
   const slotUnitAvailability = useSlotUnitAvailability({
     slotId: slotId ?? undefined,
     enabled: open && Boolean(slotId),
@@ -285,9 +287,9 @@ export function QuickBookDialog({
       })
   }, [slotUnitAvailability.data, rooms.quantities, passengers.passengers])
 
-  // Currency for voucher + payment schedule display. Defaults to EUR for the
-  // first-cut composition; follow-up wires it from the product's sell
-  // currency (usePricingPreview snapshot already carries catalog.currencyCode).
+  // Currency placeholder — used for voucher + payment schedule display.
+  // Consumers hooking in real product data should override this by wrapping
+  // the component or swapping in their own currency-aware hook.
   const currency = "EUR"
 
   const { create: createPerson } = usePersonMutation()
@@ -298,7 +300,7 @@ export function QuickBookDialog({
     setError(null)
 
     if (!product.productId) {
-      setError(messages.errorSelectProduct)
+      setError("Select a product")
       return
     }
 
@@ -306,13 +308,13 @@ export function QuickBookDialog({
     try {
       if (person.mode === "existing") {
         if (!person.personId) {
-          setError(messages.errorSelectPerson)
+          setError("Select a person or switch to create mode")
           return
         }
         resolvedPersonId = person.personId
       } else {
         if (!person.newPerson.firstName.trim() || !person.newPerson.lastName.trim()) {
-          setError(messages.errorNameRequired)
+          setError("First and last name are required")
           return
         }
         const created = await createPerson.mutateAsync({
@@ -324,19 +326,13 @@ export function QuickBookDialog({
         resolvedPersonId = created.id
       }
 
-      // Validate shared-room selection before creating the booking so we don't
-      // leave an orphan booking if the user picked "join" without a group.
       if (sharedRoom.enabled && sharedRoom.mode === "join" && !sharedRoom.groupId) {
-        setError(messages.errorSelectSharedRoom)
+        setError("Select a shared-room group to join")
         return
       }
 
       const bookingNumber = generateBookingNumber()
 
-      // Total is unknown client-side for now — the price breakdown section
-      // computes it but doesn't bubble it up. We pass null to paymentSchedule
-      // rows so "full" and "advance+balance" only emit when the operator
-      // typed amounts explicitly.
       const paymentSchedules = paymentScheduleToRows(paymentSchedule, currency, null)
 
       const travelers = passengersToRows(passengers)
@@ -377,11 +373,11 @@ export function QuickBookDialog({
         groupMembership,
       })
 
-      // Optional post-create: transition to confirmed. In operator's app.ts
-      // the notifications module has autoConfirmAndDispatch enabled, so this
-      // status change automatically triggers the doc bundle + traveler email
-      // via the booking.confirmed subscriber. If the status call fails we
-      // don't roll back — the booking exists, the operator can confirm later.
+      // Optional post-create confirm. If the app has autoConfirmAndDispatch
+      // wired on the notifications module, the status transition triggers
+      // the doc bundle + traveler email subscriber. A failed status change
+      // doesn't roll back the booking — it exists, operator can confirm
+      // manually later.
       let finalBooking = booking
       if (confirmAfterCreate) {
         try {
@@ -393,10 +389,8 @@ export function QuickBookDialog({
           setError(
             statusErr instanceof Error
               ? `Booking created but confirm failed: ${statusErr.message}`
-              : messages.errorCreateFailed,
+              : "Booking created but confirm failed",
           )
-          // Still fire onCreated so the parent closes & refreshes — the
-          // booking did land, only the confirm step tripped.
           onCreated?.(booking)
           return
         }
@@ -405,54 +399,18 @@ export function QuickBookDialog({
       onOpenChange(false)
       onCreated?.(finalBooking)
     } catch (err) {
-      setError(err instanceof Error ? err.message : messages.errorCreateFailed)
+      setError(err instanceof Error ? err.message : "Failed to create booking")
     }
   }
 
   const isSubmitting =
     quickCreateMutation.isPending || createPerson.isPending || statusMutation.isPending
 
-  const productLabels = {
-    product: messages.productLabel,
-    productSearchPlaceholder: messages.productSearchPlaceholder,
-    productSelectPlaceholder: messages.productSelectPlaceholder,
-    option: messages.optionLabel,
-    optionNone: messages.optionNone,
-  }
-
-  const personLabels = {
-    person: messages.personLabel,
-    createNewPerson: messages.createNewPerson,
-    selectExistingPerson: messages.selectExistingPerson,
-    personSearchPlaceholder: messages.personSearchPlaceholder,
-    personSelectPlaceholder: messages.personSelectPlaceholder,
-    firstName: messages.firstNameLabel,
-    firstNamePlaceholder: messages.firstNamePlaceholder,
-    lastName: messages.lastNameLabel,
-    lastNamePlaceholder: messages.lastNamePlaceholder,
-    email: messages.emailLabel,
-    emailPlaceholder: messages.emailPlaceholder,
-    phone: messages.phoneLabel,
-    phonePlaceholder: messages.phonePlaceholder,
-    organization: messages.organizationLabel,
-    organizationSearchPlaceholder: messages.organizationSearchPlaceholder,
-    organizationNone: messages.organizationNone,
-  }
-
-  const sharedRoomLabels = {
-    toggle: messages.sharedRoomLabel,
-    createMode: messages.sharedRoomCreate,
-    joinMode: messages.sharedRoomJoin,
-    selectPlaceholder: messages.sharedRoomSelectPlaceholder,
-    noGroups: messages.sharedRoomNoGroups,
-    createHint: messages.sharedRoomCreateHint,
-  }
-
   return (
     <Dialog open={open} onOpenChange={onOpenChange}>
       <DialogContent size="lg">
         <DialogHeader>
-          <DialogTitle>{messages.title}</DialogTitle>
+          <DialogTitle>Quick Book</DialogTitle>
         </DialogHeader>
         <DialogBody className="grid gap-4">
           <ProductPickerSection
@@ -460,24 +418,23 @@ export function QuickBookDialog({
             onChange={setProduct}
             enabled={open}
             lockProduct={Boolean(defaultProductId)}
-            labels={productLabels}
           />
 
           {product.productId ? (
             <div className="flex flex-col gap-1">
-              <Label>{messages.departureLabel}</Label>
+              <Label>Departure</Label>
               <Select
                 value={slotId ?? "__none__"}
                 onValueChange={(v) => setSlotId(v === "__none__" ? null : (v ?? null))}
               >
                 <SelectTrigger>
-                  <SelectValue placeholder={messages.departureSelectPlaceholder} />
+                  <SelectValue placeholder="Select a departure..." />
                 </SelectTrigger>
                 <SelectContent>
-                  <SelectItem value="__none__">{messages.departureNone}</SelectItem>
+                  <SelectItem value="__none__">No specific departure</SelectItem>
                   {slots.length === 0 ? (
                     <SelectItem value="__empty__" disabled>
-                      {messages.departureEmpty}
+                      No open departures for this product
                     </SelectItem>
                   ) : (
                     slots.map((slot) => (
@@ -495,19 +452,13 @@ export function QuickBookDialog({
             <RoomsStepperSection value={rooms} onChange={setRooms} slotId={slotId} enabled={open} />
           ) : null}
 
-          <PersonPickerSection
-            value={person}
-            onChange={setPerson}
-            enabled={open}
-            labels={personLabels}
-          />
+          <PersonPickerSection value={person} onChange={setPerson} enabled={open} />
 
           <SharedRoomSection
             value={sharedRoom}
             onChange={setSharedRoom}
             productId={product.productId || undefined}
             enabled={open}
-            labels={sharedRoomLabels}
           />
 
           {product.productId ? (
@@ -535,11 +486,11 @@ export function QuickBookDialog({
           />
 
           <div className="flex flex-col gap-2">
-            <Label>{messages.notesLabel}</Label>
+            <Label>Internal Notes</Label>
             <Textarea
               value={notes}
               onChange={(e) => setNotes(e.target.value)}
-              placeholder={messages.notesPlaceholder}
+              placeholder="Quick context for this booking..."
             />
           </div>
 
@@ -552,9 +503,13 @@ export function QuickBookDialog({
             />
             <div className="flex flex-col gap-1">
               <Label htmlFor="quickbook-confirm-after-create" className="cursor-pointer text-sm">
-                {messages.confirmAfterCreateLabel}
+                Confirm & notify traveler after creating
               </Label>
-              <p className="text-xs text-muted-foreground">{messages.confirmAfterCreateHint}</p>
+              <p className="text-xs text-muted-foreground">
+                Transitions to confirmed after create. When the notifications module's auto-dispatch
+                is on, this fires the doc bundle + traveler email via the booking.confirmed
+                subscriber.
+              </p>
             </div>
           </div>
 
@@ -568,7 +523,7 @@ export function QuickBookDialog({
             onClick={() => onOpenChange(false)}
             disabled={isSubmitting}
           >
-            {messages.cancel}
+            Cancel
           </Button>
           <Button
             type="button"
@@ -577,7 +532,7 @@ export function QuickBookDialog({
             disabled={isSubmitting || !product.productId}
           >
             {isSubmitting && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-            {messages.create}
+            Create Draft Booking
           </Button>
         </DialogFooter>
       </DialogContent>
