+++ packages/ui/registry/bookings/booking-list.tsx
@@ -6,7 +6,7 @@ import {
   formatBookingStatus,
   useBookings,
 } from "@voyantjs/bookings-react"
-import { Loader2, Plus, Search, Zap } from "lucide-react"
+import { Loader2, Plus, Search } from "lucide-react"
 import * as React from "react"
 
 import { Badge } from "@/components/ui/badge"
@@ -22,7 +22,6 @@ import {
 } from "@/components/ui/table"
 
 import { BookingDialog } from "./booking-dialog"
-import { QuickBookDialog } from "./quick-book-dialog"
 
 export interface BookingListProps {
   pageSize?: number
@@ -38,7 +37,6 @@ export function BookingList({ pageSize = 25, onSelectBooking }: BookingListProps
   const [search, setSearch] = React.useState("")
   const [offset, setOffset] = React.useState(0)
   const [dialogOpen, setDialogOpen] = React.useState(false)
-  const [quickBookOpen, setQuickBookOpen] = React.useState(false)
   const [editing, setEditing] = React.useState<BookingRecord | undefined>(undefined)
 
   const { data, isPending, isError } = useBookings({
@@ -77,10 +75,6 @@ export function BookingList({ pageSize = 25, onSelectBooking }: BookingListProps
           />
         </div>
         <div className="flex items-center gap-2">
-          <Button variant="outline" onClick={() => setQuickBookOpen(true)}>
-            <Zap className="mr-2 size-4" />
-            Quick Book
-          </Button>
           <Button
             onClick={() => {
               setEditing(undefined)
@@ -180,14 +174,6 @@ export function BookingList({ pageSize = 25, onSelectBooking }: BookingListProps
         onOpenChange={setDialogOpen}
         booking={editing}
         onSuccess={(booking) => {
-          onSelectBooking?.(booking)
-        }}
-      />
-
-      <QuickBookDialog
-        open={quickBookOpen}
-        onOpenChange={setQuickBookOpen}
-        onCreated={(booking) => {
           onSelectBooking?.(booking)
         }}
       />
